feat: enable binary copy in operation, add more tests, rename test classes
