+++ postgresql/src/main/kotlin/io/github/clasicrando/kdbc/postgresql/copy/PgBinaryCopyRow.kt
@@ -1,0 +1,8 @@
+package io.github.clasicrando.kdbc.postgresql.copy
+
+import io.github.clasicrando.kdbc.postgresql.statement.PgEncodeBuffer
+
+interface PgBinaryCopyRow {
+    val valueCount: Short
+    fun encodeValues(buffer: PgEncodeBuffer)
+}
