+++ postgresql/src/main/kotlin/io/github/clasicrando/kdbc/postgresql/copy/PgCsvCopyRow.kt
@@ -1,0 +1,5 @@
+package io.github.clasicrando.kdbc.postgresql.copy
+
+interface PgCsvCopyRow {
+    val values: List<Any?>
+}
